Bridge native widths and strengthen cross-ABI tests
Translate CK_ULONG values and nested templates across 32-bit, 64-bit, and Windows layouts. Expand live and mock tests, then split large conversion and test modules.

+++ crates/server/src/server/grpc_service/combined/sign_encrypt.rs
@@ -20,7 +20,7 @@ async fn resolve_backend_session(
     match backend_session {
         None => Err(CkRv::CRYPTOKI_NOT_INITIALIZED),
         Some(None) => Err(CkRv::SESSION_HANDLE_INVALID),
-        Some(Some(handle)) => Ok(CkSessionHandle(handle.0)),
+        Some(Some(handle)) => Ok(CkSessionHandle(handle.0 as u64)),
     }
 }
